pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-as-ajudge-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langgraph>=0.6.4",
    "tomli>=2.2.1",
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.3.29",
    "langsmith>=0.2.3",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "streamlit>=1.28.0",
    "agno>=1.7.9",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pip>=25.2",
    "ipykernel>=6.30.1",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["laaj*"]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "tomli>=2.0.1",
]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
